pyproject.toml
[tool.poetry]
name = "tabulator"
version = "0.1.0"
description = ""
authors = ["samuel baetz"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
djangorestframework = "^3.12.3"
whitenoise = "^5.2.0"
gunicorn = "^20.0.4"
django-environ = "^0.4.5"
psycopg2 = "^2.8.6"
Pillow = "^8.1.2"
django-filter = "^2.4.0"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
